command-line-params move servant API type to API module to break Common dependency on Server
