+++ app/astro/src/Astro/Server.hs
@@ -6,7 +6,6 @@
 
 module Astro.Server
   ( runAstroServer
-  , astroAPI
   ) where
 
 import           Control.Monad
@@ -38,57 +37,13 @@ import           Astro.Catalogue
 import           Astro.Types
 
 
-type AstroAPI =
-  (  "meteors"
-    :> QueryParam "mass" Int
-    :> QueryParam "size" Int
-    :> Get '[JSON] Meteors
-    )
-  :<|>
-    (  "meteor"
-    :> ReqBody '[JSON] API.MeteorTemplate
-    :> Post '[JSON] MeteorId
-    )
-  :<|>
-    (  "asteroid"
-    :> ReqBody '[JSON] API.AsteroidTemplate
-    :> Post '[JSON] AsteroidId
-    )
-  :<|>
-    (  "object_template"                                  -- route POST "/object_template"
-    :> ReqBody '[JSON] API.AstroObjectTemplate
-    :> Post '[JSON] AstroObjectId
-    )
-  :<|>
-    "object" :>
-    (
-       ( Capture "object_id" AstroObjectId                -- route GET "/object"
-       :> Get '[JSON] (Maybe AstroObject)
-       )
-     :<|>
-       ( "orbital"                                        -- route POST "/object/orbital"
-       :> Capture "object_id" AstroObjectId
-       :> ReqBody '[JSON] Orbital
-       :> Post '[JSON] AstroObjectId
-       )
-     :<|>
-       ( "physical"                                       -- route POST "/object/physical
-       :> Capture "object_id" AstroObjectId
-       :> ReqBody '[JSON] Physical
-       :> Post '[JSON] AstroObjectId
-       )
-    )
-
-
-astroAPI :: Proxy AstroAPI
-astroAPI = Proxy
 
 data Env = Env !R.AppRuntime !AppState
 type AppHandler = ReaderT Env (ExceptT ServerError IO)
-type AppServer = ServerT AstroAPI AppHandler
+type AppServer = ServerT API.AstroAPI AppHandler
 
-astroServer :: Env -> Server AstroAPI
-astroServer env = hoistServer astroAPI f astroServer'
+astroServer :: Env -> Server API.AstroAPI
+astroServer env = hoistServer API.astroAPI f astroServer'
   where
     f :: ReaderT Env (ExceptT ServerError IO) a -> Handler a
     f r = do
@@ -98,7 +53,7 @@ astroServer env = hoistServer astroAPI f astroServer'
         Right res -> pure res
 
 astroBackendApp :: Env -> Application
-astroBackendApp = serve astroAPI . astroServer
+astroBackendApp = serve API.astroAPI . astroServer
 
 runApp :: L.AppL a -> AppHandler a
 runApp flow = do
